# api-docs.yaml
openapi: 3.1.0
info:
  title: ReelRAG – Video Comparison API
  description: |
    RAG-powered API for comparing YouTube and Instagram videos.
    Extracts transcripts, calculates engagement metrics, stores embeddings
    in ChromaDB, and exposes a streaming chat interface backed by LangGraph.
  version: 1.0.0
  contact:
    name: ReelRAG Team
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/videos/{video_id}:
    get:
      tags: [videos]
      summary: Get metadata for a specific video
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            enum: [A, B]
      responses:
        "200":
          description: Video metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoMetadata"
        "404":
          description: Video not found
...
components:
  schemas:
...
    VideoMetadata:
      type: object
      properties:
        video_id:
          type: string
          enum: [A, B]
        title:
          type: string
        creator_name:
          type: string
        follower_count:
          type: integer
          nullable: true
        views:
          type: integer
          nullable: true
        likes:
          type: integer
          nullable: true
        comments:
          type: integer
          nullable: true
        upload_date:
          type: string
          nullable: true
        duration_seconds:
          type: integer
          nullable: true
        hashtags:
          type: array
          items:
            type: string
        transcript:
          type: string
          nullable: true
        engagement_rate:
          type: number
          format: float
          nullable: true
          description: "(likes + comments) / views * 100"
        source:
          type: string
          enum: [youtube, instagram]
        url:
          type: string
...